This partially addresses NCL-696.
I'm checking it in now so we can have latest versions of ispan,
fspan, num, and ind.

NclIfbindirread - this function was somewhat complicated, but mainly,
I just needed to make sure the dimsizes array was approriately coerced
to int or long.

_NclIisbigendian - no change

_NclIcbinread - same situation as _NclIfbindirread

_NclIfbinnumrec - no change, already had dimsizes converted

_NclIfbinrecwrite - no change, already had dimsizes converted

_NclIfbinrecread - same situation as _NclIfbindirread, a little strange
because not consistently done (variable called "dimsizes" instead of 
"dimensions", which I changed)

_NclIfbinread - same as _NclIfbinrecread.

_NclIasciiwrite - no changes

_NclIasciiread - same as _NclIfbinrecread, had to convert input dimension
sizes to ng_size_t.

_NclIchngdir - no change, not even advertised?

No changes to these routines:

_Nclfbindirwrite
_Nclcbinwrite
_Nclfbinwrite
_NclIsleep
_NclIprompt
_NclIrand
_NclIsrand
_NclIabs
_NclIncargversion
_NclIncargpath
_NclIgetenv
_NclIushorttoint
_NclIinttoshort
_NclIinttobyte
_NclIinttochar
_NclIchartoint
_NclIshorttobyte
_NclIshorttochar
_NclIchartoshort
_NclIlongtoint
_NclIlongtoshort
_NclIlongtobyte
_NclIlongtochar
_NclIchartolong
_NclIfloattoshort
_NclIfloattoint
_NclIfloattolong
_NclIfloattobyte
_NclIfloattochar
_NclIchartofloat
_NclIdoubletobyte
_NclIdoubletochar
_NclIchartodouble
_NclIdoubletoshort
_NclIdoubletoint
_NclIdoubletolong
_NclIdoubletofloat
_NclIstringtolong
_NclIstringtoulong
_NclIstringtoint64
_NclIstringtouint64
_NclIstringtoshort
_NclIstringtoushort
_NclIstringtointeger
_NclIstringtouint
_NclIstringtodouble
_NclIstringtofloat
_NclIstringtochar
_NclIchartostring
_NclIIsProc
_NclIExit
_NclIStatusExit
_NclIIsFunc
_NclIUnDef
_NclIIsDefined
_NclIIsVar
_NclIIsCoord
_NclIIsAtt
_NclIIsDim
_NclIIsDimNamed
_NclIIsFileVar
_NclIIsFileVarAtt
_NclIIsFileVarCoord
_NclIIsFileVarDim

_Ncl1dtond - dimension sizes had to be handled correctly.
_Nclndto1d - no changes, there are no dimension sizes here

_Nclnum - Now returns int or long, depending on NG32BIT.

_Nclind - same kind of deal as _Nclnum, except now returning an array of
ints or longs. This one involved a bit of work since there are four main
"if/else" blocks that had to be changed.

_Nclispan - This function was updated to allow multiple kinds of input:

 - Only byte, int8, short, int, int64, long values can be input.
 - Int64 will be returned if any int64 are input.
 - Long will be returned if any long (and no int64) are input.
 - Int will be returned if byte, int8, short, or int are input (and no 
   long or int64)
 - An error will be issued if any other type input.
 - You don't need to check for INT32_MAX because whatever type 
   comes in will be already dealing with this. That is, if a user
   wants ispan to have numbers > INT32_MAX, then 64-bit ints (or
   longs on a 64-bit machine) will have to have been input.

_Nclfspan - small change to allow int8 and int64 input for 3rd argument.
    Had to make change in AddBuiltIns.c for snumeric type for 3rd
    argument.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11821 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/AddIntrinsics.c
@@ -559,7 +559,7 @@ void _NclAddIntrinsics
 	args[0].dim_sizes[0] = 1;
 	args[0].is_dimsizes = 1;
 	args[0].n_dims = 1;
-	args[1].arg_data_type = _NclLookUp("long");
+	args[1].arg_data_type = _NclLookUp("numeric");
 	args[1].is_dimsizes = 0;
 	args[1].n_dims = 1;
 	args[2].arg_data_type = _NclLookUp("string");
@@ -577,7 +577,7 @@ void _NclAddIntrinsics
 	args[1].dim_sizes[0] = 1;
 	args[1].is_dimsizes = 1;
 	args[1].n_dims = 1;
-	args[2].arg_data_type = _NclLookUp("long");
+	args[2].arg_data_type = _NclLookUp("numeric");
 	args[2].is_dimsizes = 1;
 	args[2].n_dims = 1;
 	args[2].dim_sizes[0] = -1;
@@ -596,7 +596,7 @@ void _NclAddIntrinsics
 	args[1].dim_sizes[0] = 1;
 	args[1].is_dimsizes = 1;
 	args[1].n_dims = 1;
-	args[2].arg_data_type = _NclLookUp("long");
+	args[2].arg_data_type = _NclLookUp("numeric");
 	args[2].is_dimsizes = 1;
 	args[2].n_dims = 1;
 	args[2].dim_sizes[0] = -1;
@@ -657,7 +657,7 @@ void _NclAddIntrinsics
 	args[0].dim_sizes[0] = 1;
 	args[0].is_dimsizes = 1;
 	args[0].n_dims = 1;
-	args[1].arg_data_type = _NclLookUp("long");
+	args[1].arg_data_type = _NclLookUp("numeric");
 	args[1].is_dimsizes = 0;
 	args[1].n_dims = 1;
 	args[2].arg_data_type = _NclLookUp("string");
@@ -671,7 +671,7 @@ void _NclAddIntrinsics
 	args[0].dim_sizes[0] = 1;
 	args[0].is_dimsizes = 1;
 	args[0].n_dims = 1;
-	args[1].arg_data_type = _NclLookUp("long");
+	args[1].arg_data_type = _NclLookUp("numeric");
 	args[1].is_dimsizes = 0;
 	args[1].n_dims = 1;
 	args[2].arg_data_type = _NclLookUp("string");
